+++ retroauth-core/src/main/java/com/andretietz/retroauth/TokenStorage.java
@@ -22,14 +22,6 @@ package com.andretietz.retroauth;
 public interface TokenStorage<OWNER, TOKEN_TYPE, TOKEN> {
 
     /**
-     * Creates a token type object. This doe
-     *
-     * @param annotationValues The values from the {@link Authenticated} annotation
-     * @return a token type.
-     */
-    TOKEN_TYPE createType(int[] annotationValues);
-
-    /**
      * This method returns an authentication token. If there's no token, you should try
      * authenticating your user.
      *